Move everything to a package
