+++ ember-dev.yml
@@ -15,3 +15,4 @@ testing_suites:
 testing_packages:
   - ember-data
   - ember-inflector
+  - activemodel-adapter
