+++ packages/activemodel-adapter/lib/main.js
@@ -1,0 +1,1 @@
+require('activemodel-adapter/system');
