+++ packages/activemodel-adapter/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "activemodel-adapter",
+  "summary": "ActiveModelAdapter for Ember Data",
+  "description": "",
+  "homepage": "http://github.com/emberjs/ember.js",
+  "author": ["Bradley Priest"],
+  "version": "1.0.0-beta.1",
+
+  "directories":{
+    "lib": "lib"
+  },
+
+  "dependencies": {
+    "ember-runtime": "1.0.0",
+    "ember-data": "1.0.0"
+  },
+
+  "bpm:build": {
+    "bpm_lib.js": {
+      "files": ["lib"],
+      "modes": "*"
+    }
+  }
+}
