+++ packages/ember-data/lib/adapters.js
@@ -4,4 +4,3 @@
 
 require("ember-data/adapters/fixture_adapter");
 require("ember-data/adapters/rest_adapter");
-require("ember-data/adapters/active_model_adapter");
